Move files and rename package from c2c.bottomup to it.unibro.c2c
Package name follows sandard java conventions:

https://docs.oracle.com/javase/tutorial/java/package/namingpkgs.html
https://www.oracle.com/java/technologies/javase/codeconventions-namingconventions.html
